+++ @geogebra/solver-sdk/src/parser/tree-to-solver.ts
@@ -53,6 +53,8 @@ export function treeToSolver(n: ExpressionTree): string {
       return dec(`sqrt[${rec(n.args[0])}]`);
     case 'Root':
       return dec(`root[${rec(n.args[0])}, ${rec(n.args[1])}]`);
+    case 'AbsoluteValue':
+      return dec(`abs[${rec(n.args[0])}]`);
     case 'Equation':
       return dec(`${rec(n.args[0])} = ${rec(n.args[1])}`);
     case 'EquationSystem':