+++ CHANGELOG.md
@@ -14,6 +14,7 @@ contents to a new release section (e.g. _[0.x]_).
 - Added CHANGELOG.md (this file) (PLUT-555).
 - Ability to collect example inputs for a given explanation key (PLUT-569)
 - All requests for a solution are logged, together with solution and timing (PLUT-548)
+- Added support for the absolute value operator and rules for resolving it in simple cases (PLUT-557)
 
 ### Fixed